+++ routers/sign-in.js
@@ -1,10 +1,10 @@
 const path = require("path");
 const fs = require("fs");
-const dbHandler = require("../api/db-handler");
+const fileHandler = require("../api/file-handler");
 
 const getSignInPage = () => async (req, res, next) => {
   const publicPath = path.join(__dirname, "../public");
-  const data = await dbHandler.readFile(`${publicPath}/sign-in.html`);
+  const data = await fileHandler.readFile(`${publicPath}/sign-in.html`);
 
   res.writeHead(200, { "Content-Type": "text/html" });
   res.end(data);
